Sort out folder and python files
