+++ rake_tasks/crazy_fun/mappings/file_copy_hack.rb
