+++ rake_tasks/crazy_fun/mappings/python/add_normal_dependencies.rb
@@ -1,0 +1,8 @@
+module Python
+  class AddNormalDependencies < PyTask
+    def handle(fun, dir, args)
+      target = Rake::Task[task_name(dir, "#{args[:name]}")]
+      add_dependencies(target, dir, args[:deps])
+    end
+  end
+end
